Common parts of notify-client isolated
