+++ pkg/virt-launcher/notify-client/BUILD.bazel
@@ -2,7 +2,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["client.go"],
+    srcs = [
+        "client.go",
+        "client-libvirt.go",
+    ],
     importpath = "kubevirt.io/kubevirt/pkg/virt-launcher/notify-client",
     visibility = ["//visibility:public"],
     deps = [
